% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typojoin.R
\name{typo_inner_join}
\alias{typo_inner_join}
\title{Deterministic Typo-Tolerant Join}
\usage{
typo_inner_join(
  x,
  y,
  by_x,
  by_y = by_x,
  max_dist = 2,
  method = "lv",
  keep_y_key = TRUE,
  assert_complete = FALSE
)
}
\arguments{
\item{x, y}{Data frames to join.}

\item{by_x}{A character string giving the column name in \code{x}
to use as the join key.}

\item{by_y}{A character string giving the column name in \code{y}
to use as the join key. Defaults to \code{by_x}.}

\item{max_dist}{Maximum string distance.}

\item{method}{String distance method.}

\item{keep_y_key}{Keep messy key in output.}

\item{assert_complete}{Error if unmatched rows remain.}
}
\value{
A joined data frame.
}
\description{
Reconciles two data frames that should match except for
minor formatting differences or small alphabetical typos.
}
